+++ package.json
@@ -10,6 +10,8 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@aries-framework/core": "^0.2.5",
+    "@aries-framework/react-hooks": "^0.3.1",
     "@expo/ngrok": "^2.5.0",
     "@metamask/bip39": "^4.0.0",
     "axios": "^1.1.3",
@@ -24,6 +26,7 @@
     "hdkey": "^2.0.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-i18next": "^12.0.0",
     "react-native": "0.68.2",
     "react-native-fs": "^2.20.0",
     "react-native-web": "0.17.7"
